Splitting the legacy GUI and its dependency on PyQt5 into a separate package.
