+++ src/legacy/python/legacy_gui/README.md
@@ -1,0 +1,6 @@
+# RLBot Legacy GUI
+This is a very old GUI, still kept around on the off chance
+that someone is still relying on it in their workflow.
+
+See https://github.com/RLBot/RLBotGUI for our new user
+interface under active development.
