+++ src/legacy/python/legacy_gui/publish-to-pypi-prod.bat
@@ -1,0 +1,1 @@
+python setup.py sdist upload
