+++ src/legacy/python/legacy_gui/publish-to-pypi-test.bat
@@ -1,0 +1,1 @@
+python setup.py sdist upload -r pypitest
