+++ src/legacy/python/legacy_gui/rlbot_legacy_gui/rocket_league_items.json
